+++ src/main/java/axoloti/realunits/LinRatio.java
@@ -49,7 +49,7 @@ public class LinRatio implements NativeToReal {
             double num;
 
             try {
-                num = Float.parseFloat(matcher.group("num"));
+                num = Double.parseDouble(matcher.group("num"));
             } catch (java.lang.NumberFormatException ex) {
                 throw new ParseException("Not LinRatio", 0);
             }
