+++ src/main/java/axoloti/realunits/SquareRatio.java
@@ -50,7 +50,7 @@ public class SquareRatio implements NativeToReal {
             double num;
 
             try {
-                num = Float.parseFloat(matcher.group("num"));
+                num = Double.parseDouble(matcher.group("num"));
             } catch (java.lang.NumberFormatException ex) {
                 throw new ParseException("Not PitchToRatio", 0);
             }
